+++ blockchain/shardutils.go
@@ -3,14 +3,15 @@ package blockchain
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/incognitochain/incognito-chain/incognitokey"
 	"reflect"
 	"sort"
 	"strconv"
 	"strings"
 
 	"github.com/incognitochain/incognito-chain/common"
+	"github.com/incognitochain/incognito-chain/common/base58"
 	"github.com/incognitochain/incognito-chain/database"
+	"github.com/incognitochain/incognito-chain/incognitokey"
 	"github.com/incognitochain/incognito-chain/metadata"
 	"github.com/incognitochain/incognito-chain/privacy"
 	"github.com/incognitochain/incognito-chain/transaction"
@@ -248,6 +249,48 @@ func checkReturnStakingTxExistence(txId string, shardBlock *ShardBlock) bool {
 		}
 	}
 	return false
+}
+
+func getRequesterFromPKnCoinID(pk privacy.PublicKey, coinID common.Hash) string {
+	requester := base58.Base58Check{}.Encode(pk, common.Base58Version)
+	return fmt.Sprintf("%s-%s", requester, coinID.String())
+}
+
+func reqTableFromReqTxs(
+	transactions []metadata.Transaction,
+) map[string]metadata.Transaction {
+	txRequestTable := map[string]metadata.Transaction{}
+	for _, tx := range transactions {
+		if tx.GetMetadataType() == metadata.WithDrawRewardRequestMeta {
+			requestMeta := tx.GetMetadata().(*metadata.WithDrawRewardRequest)
+			key := getRequesterFromPKnCoinID(requestMeta.PaymentAddress.Pk, requestMeta.TokenID)
+			txRequestTable[key] = tx
+		}
+	}
+	return txRequestTable
+}
+
+func filterReqTxs(
+	transactions []metadata.Transaction,
+	txRequestTable map[string]metadata.Transaction,
+) []metadata.Transaction {
+	res := []metadata.Transaction{}
+	for _, tx := range transactions {
+		if tx.GetMetadataType() == metadata.WithDrawRewardRequestMeta {
+			requestMeta := tx.GetMetadata().(*metadata.WithDrawRewardRequest)
+			key := getRequesterFromPKnCoinID(requestMeta.PaymentAddress.Pk, requestMeta.TokenID)
+			txReq, ok := txRequestTable[key]
+			if !ok {
+				continue
+			}
+			cmp, err := txReq.Hash().Cmp(tx.Hash())
+			if (err != nil) || (cmp != 0) {
+				continue
+			}
+		}
+		res = append(res, tx)
+	}
+	return res
 }
 
 //=======================================END SHARD BLOCK UTIL
